Updated SceneFileParser in order to manage UBOs.
Updated the CastorViewer GUI.
Using wxAuiManager, to have logs and properties dock panels.
Now modifying PassPanel, to be able to display it in a reduced window.
Corrected the scene reload, Castor3D don't crash anymore because of that.

Few corrections to build in release

Made it build on Linux/GCC.
Now I have a special panel for properties.

Reviewing the wxGeometriesListFrame class.
It displays all scene objects, so a rename will be necessary.
Now need to implement properties for each object type, and it should be good for this panel.
I will have to be a bit smarter in my way to handle the properties panels...

Properties are nearly finished in read only, just remains viewport.

Nearly all Overlay properties are modifiable, just remains Font property for text overlays.

Repaired shaders and buffers filling on Intel GPUs.